Исправил ряд функций до стандарта 17
- переехал с size_t на int
- переделал работу с директориями
- переделал работу с классами символов (см. ParserBackground)

+++ SystemGeneration/AffineTransformation.cpp
@@ -10,7 +10,7 @@ using namespace transformations;
 
 AffineTransformation::AffineTransformation(const MatrixB& M, const RowB& v, bool print_or_not)
 {
-	size_t n = M.size();
+	int n = M.size();
 	if (v.size() != n)
 		throw exception("Matrix M and vector v for affine transformation have same dimension");
 
@@ -18,10 +18,10 @@ AffineTransformation::AffineTransformation(const MatrixB& M, const RowB& v, bool
 	RowB* cur_vec;
 	Polynomial cur;
 	size_t prev_num = 0;
-	for (size_t i = 0; i < n; i++)
+	for (int i = 0; i < n; i++)
 	{
 		cur_vec = M[i];
-		for (size_t j = 0; j < n; j++)
+		for (int j = 0; j < n; j++)
 		{
 			if (cur_vec->get(j) == TRUE)
 			{
@@ -39,13 +39,13 @@ AffineTransformation::AffineTransformation(const MatrixB& M, const RowB& v, bool
 
 		if (i % 50 == 0 && print_or_not)
 		{
-			for (size_t i = 0; i < prev_num; i++) cout << '\b';
+			for (int i = 0; i < prev_num; i++) cout << '\b';
 			cout << i << '/' << n;
 			prev_num = to_string(i).length() + to_string(n).length() + 1;
 		}
 
 	}
-	for (size_t i = 0; i < prev_num; i++) cout << '\b';
+	for (int i = 0; i < prev_num; i++) cout << '\b';
 
 }
