+++ SystemGeneration/Matrix.cpp
@@ -7,7 +7,7 @@ using namespace matrixes;
 
 
 template <typename _T>
-BOOL Matrix<_T>::init_zeros(size_t dimension, size_t dimension2)
+BOOL Matrix<_T>::init_zeros(int dimension, int dimension2)
 {
 	data.clear();
 	data.reserve(dimension);
@@ -15,7 +15,7 @@ BOOL Matrix<_T>::init_zeros(size_t dimension, size_t dimension2)
 	Row <_T> * row;
 	data = std::vector<Row <_T> *>(dimension);
 
-	for (size_t i = 0; i < dimension; i++)
+	for (int i = 0; i < dimension; i++)
 	{
 		try {
 			row = new Row<_T>(dimension2);
@@ -37,7 +37,7 @@ BOOL Matrix<_T>::init_zeros(size_t dimension, size_t dimension2)
 }
 
 template <typename _T>
-inline void Matrix<_T>::swap(size_t index1, size_t index2)
+inline void Matrix<_T>::swap(int index1, int index2)
 {
 	if (index1 != index2)
 	{
@@ -49,20 +49,20 @@ inline void Matrix<_T>::swap(size_t index1, size_t index2)
 template <typename _T>
 void Matrix<_T>::initInverse(const Matrix<_T>& matrix)
 {
-	size_t n = matrix.size();
+	int n = matrix.size();
 	if (n == 0) return;
 
 	// слева исходная матрица, справа единичная. По окончании работы метода слева будет единичная матрица, а справа обратная к исходной, что и требуется
 	init_zeros(n, 2 * n);
-	for (size_t i = 0; i < n; i++)
+	for (int i = 0; i < n; i++)
 	{
-		for (size_t j = 0; j < n; j++)
+		for (int j = 0; j < n; j++)
 			get(i)->set(j, matrix[i]->get(j));
 		get(i)->set(i + n, 1);
 	}
 
 	// построение верхнетреугольной матрицы
-	size_t row_num, offset = 0, row_num_top = 0;
+	int row_num, offset = 0, row_num_top = 0;
 	Row<_T>* r0;
 	while (offset < n - 1)
 	{
@@ -81,7 +81,7 @@ void Matrix<_T>::initInverse(const Matrix<_T>& matrix)
 
 		// для каждой строки ниже, если первый элемент 1, заменяем строку ее сумму с первой строкой. В итоге во всех строках ниже стоят нули
 		r0 = data[offset];
-		for (size_t i = offset + 1; i < n; i++)
+		for (int i = offset + 1; i < n; i++)
 		{
 			if (get(i)->get(offset) == TRUE)
 				get(i)->operator^=(*r0);
@@ -110,7 +110,7 @@ void Matrix<_T>::initInverse(const Matrix<_T>& matrix)
 	}
 
 	// взятие "правой половины" матрицы
-	for (size_t i = 0; i < n; i++)
+	for (int i = 0; i < n; i++)
 	{
 		*get(i) = Row<_T>(vector<_T>(get(i)->begin() + n, get(i)->end()));
 	}
@@ -121,18 +121,18 @@ template <typename _T>
 void Matrix<_T>::print() const
 {
 	if (!data.empty()) {
-		size_t n = size();
-		size_t n2 = data[0]->size(); // в процессе построения обратной матрицы n может быть не равно n2
+		int n = size();
+		int n2 = data[0]->size(); // в процессе построения обратной матрицы n может быть не равно n2
 
 		cout << endl << "      ";
-		for (size_t i = 0; i < n; i++)
+		for (int i = 0; i < n; i++)
 			cout << i << "  ";
 		cout << endl << endl;
 
-		for (size_t i = 0; i < n; i++)
+		for (int i = 0; i < n; i++)
 		{
 			cout << "  " << i << "   ";
-			for (size_t j = 0; j < n2; j++)
+			for (int j = 0; j < n2; j++)
 				cout << data[i]->get(j) << "  ";
 			cout << endl;
 		}
