+++ SystemGeneration/Monomial.h
@@ -15,25 +15,25 @@ namespace polynomials {
 		// Список номеров индексов переменных терма, отсортированный по возрастанию
 		// Очевидно, в терме каждая переменная встречается не более, чем 1 раз
 		// Например, для монома х1х4х2 vars={1,2,4}
-		std::vector<size_t> vars;
+		std::vector<int> vars;
 		void simplify();
-		size_t n_max; // наибольший номер присутствующих переменных
+		int n_max; // наибольший номер присутствующих переменных
 
 	public:
 
-		Monomial(std::vector<size_t>& in_grades) : n_max(0)
+		Monomial(std::vector<int>& in_grades) : n_max(0)
 		{
 			vars = std::move(in_grades);
 			simplify();
 		}
 		Monomial() : n_max(0) {};
-		Monomial(size_t num) : n_max(0)
+		Monomial(int num) : n_max(0)
 		{
 			vars.push_back(num);
 			simplify();
 		}
 
-		size_t get_n_max() const
+		int get_n_max() const
 		{
 			return n_max;
 		}
@@ -41,21 +41,21 @@ namespace polynomials {
 		BOOL substitute(const std::vector<BOOL>& values) const;
 		void toString(std::string& out) const;
 
-		inline size_t size() const
+		inline int size() const
 		{
-			return vars.size();
+			return (int)vars.size();
 		}
 
 		// получение (но не изменение) номера i-ой переменной монома
-		inline size_t operator[](size_t i) const
+		inline int operator[](int i) const
 		{
 			return vars[i];
 		}
 
 		void operator*= (const Monomial& m2)
 		{
-			size_t sz = m2.size();
-			for (size_t i = 0; i < sz; i++)
+			int sz = m2.size();
+			for (int i = 0; i < sz; i++)
 				vars.push_back(m2[i]);
 
 			simplify();
